% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{accessibility_score}
\alias{accessibility_score}
\title{Estimates the unranked Accesibility score for TFBS}
\usage{
accessibility_score(data = "", output_dir = "", plot = TRUE, verbose = FALSE)
}
\arguments{
\item{data}{DATA.FRAME with data or Path to TXT file}

\item{output_dir}{Directory to output results. If not provided then outputs in current directory}

\item{plot}{Create a plot with frequency data. Default TRUE.}

\item{verbose}{Enables progress messages. Default FALSE}
}
\description{
This function takes a DATA.FRAME/ TXT file with the TFBS coverage data and
estimates the Accesibility score as the range between the global maximum and minimum of
the high-frequency signal. In order, to estimate the high-frequency signal, and remove
the local biases the Savitzky-Golay filter is used.
}
