% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze_MR_tfs}
\alias{analyze_MR_tfs}
\title{Analyze methylation ration around TFBS}
\usage{
analyze_MR_tfs(
  bin_path = "tools/samtools/samtools",
  bin_path2 = "tools/PileOMeth/output/MethylDackel",
  bed_dir = "",
  tfs = "",
  bam = "",
  tfbs_start = 1000,
  tfbs_end = 1000,
  max_regions = 1e+05,
  mapq = 10,
  phred = 5,
  verbose = FALSE,
  output_dir = "",
  plot = TRUE,
  keep_strand = TRUE,
  bin_width = 50
)
}
\arguments{
\item{bin_path}{Path to binary. Default tools/bedtools2/bin/bedtools}

\item{bin_path2}{Path to secondary binary. Default tools/samtools/samtools}

\item{bed_dir}{Path to directory with BED files for TFBS}

\item{tfs}{Number of TFs to analyze or a list with TFs to analyze. Default none will analyze all TFs in BED directory.}

\item{bam}{Path to BAM file}

\item{tfbs_start}{Number of bases to analyze forward from TFBS central point. Default 1000}

\item{tfbs_end}{Number of bases to analyze  backward from TFBS central point. Default 1000}

\item{max_regions}{Max number of TFBS to analyze. Default 100000}

\item{mapq}{Min quality of mapping reads. Default 0}

\item{verbose}{Enables progress messages. Default FALSE}

\item{output_dir}{Directory to output results. If not provided then outputs in current directory}

\item{plot}{Create plots. Default TRUE.}

\item{mean_cov}{Mean genome wide coverage. If not provided it will be estimated.}

\item{norm}{Path to TXT file with normalized local coverage}

\item{cov_limit}{Max base depth. Default 1000}

\item{threads}{Number of threads. Default 1}
}
\value{
A DATA.FRAME with coverage data
}
\description{
This function takes the path to the directory with BED files for each TF, a BAM file with an aligned bisulfite treated sequence
and estimates methylation ratio values around TFBS.
}
