% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze_tfs}
\alias{analyze_tfs}
\title{Analyze TFs for a sample}
\usage{
analyze_tfs(
  bin_path = "tools/bedtools2/bin/bedtools",
  bin_path2 = "tools/samtools/samtools",
  bed_dir = "",
  tfs = "",
  bam = "",
  tfbs_start = 1000,
  tfbs_end = 1000,
  mean_cov = "",
  norm = "",
  threads = 1,
  cov_limit = 1000,
  max_regions = 1e+05,
  mapq = 0,
  verbose = FALSE,
  output_dir = "",
  plot = TRUE,
  tfbs_data_keep = TRUE
)
}
\arguments{
\item{bin_path}{Path to binary. Default tools/bedtools2/bin/bedtools}

\item{bin_path2}{Path to secondary binary. Default tools/samtools/samtools}

\item{bed_dir}{Path to directory with BED files for TFBS}

\item{tfs}{Number of TFs to analyze or a list with TFs to analyze. Default none will analyze all TFs in BED directory.}

\item{bam}{Path to BAM file}

\item{tfbs_start}{Number of bases to analyze forward from TFBS central point. Default 1000}

\item{tfbs_end}{Number of bases to analyze  backward from TFBS central point. Default 1000}

\item{mean_cov}{Mean genome wide coverage. If not provided it will be estimated.}

\item{norm}{Path to TXT file with normalized local coverage}

\item{threads}{Number of threads. Default 1}

\item{cov_limit}{Max base depth. Default 1000}

\item{max_regions}{Max number of TFBS to analyze. Default 100000}

\item{mapq}{Min quality of mapping reads. Default 0}

\item{verbose}{Enables progress messages. Default FALSE}

\item{output_dir}{Directory to output results. If not provided then outputs in current directory}

\item{plot}{Create plots. Default TRUE.}

\item{tfbs_data_keep}{Keep unprocessed coverage tfbs data. Default FALSE.}
}
\value{
A DATA.FRAME with coverage data
}
\description{
This function takes the path to the directory with BED files for each TF, a BAM file with an aligned sequence and a TXT file with the normalized
local coverage values for CNA, and estimates the corrected mean depth coverage values around TFBS, as well as the accessibilty scores for them.
For better understanding check: https://www.nature.com/articles/s41467-019-12714-4
}
