% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze_tss_around_position}
\alias{analyze_tss_around_position}
\title{Analyze Mean Coverage Depth around TFBS}
\usage{
analyze_tss_around_position(
  bin_path = "tools/bedtools2/bin/bedtools",
  bin_path2 = "tools/samtools/samtools",
  bed = "",
  bam = "",
  tss_start = 1000,
  tss_end = 1000,
  mean_cov = "",
  norm = "",
  threads = 1,
  cov_limit = 1000,
  max_regions = 1e+05,
  mapq = 0,
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{Path to binary. Default tools/bedtools2/bin/bedtools}

\item{bin_path2}{Path to secondary binary. Default tools/samtools/samtools}

\item{bed}{Path to BED file.}

\item{bam}{Path to BAM file.}

\item{tss_start}{Number of bases to analyze forward from TFBS central point. Default 1000}

\item{tss_end}{Number of bases to analyze  backward from TFBS central point. Default 1000}

\item{mean_cov}{Mean genome wide coverage. If not provided it will be estimated.}

\item{threads}{Number of threads. Default 1}

\item{cov_limit}{Max base depth. Default 1000}

\item{max_regions}{Max number of TFBS to analyze. Default 100000}

\item{mapq}{Min quality of mapping reads. Default 0}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function takes a BED file with TFBS, a BAM file with an aligned sequence and a TXT file with the normalized
local coverage values for CNA, and outputs the corrected mean depth coverage values around TFBS, as well as
generates a pdf with a plot for them. For better understanding check: https://www.nature.com/articles/s41467-019-12714-4
}
