% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{calculate_coverage_tfbs}
\alias{calculate_coverage_tfbs}
\title{Calculate Mean Coverage Depth around TFBS}
\usage{
calculate_coverage_tfbs(
  bin_path = "tools/samtools/samtools",
  ref_data = "",
  bam = "",
  sample_name = "",
  tf_name = "",
  mean_cov = "",
  norm_log2 = "",
  tfbs_start = 1000,
  tfbs_end = 1000,
  cov_limit = 1000,
  mapq = 0,
  threads = 1,
  output_dir = ""
)
}
\arguments{
\item{bin_path}{Path to binary. Default tools/samtools/samtools}

\item{ref_data}{Data from BED file as Data.frame}

\item{bam}{Path to BAM file.}

\item{sample_name}{Sample name}

\item{tf_name}{Transcription Factor name}

\item{mean_cov}{Mean genome wide coverage.}

\item{norm_log2}{Data.frame with normalized local coverage}

\item{tfbs_start}{Number of bases to analyze forward from TFBS central point. Default 1000}

\item{tfbs_end}{Number of bases to analyze  backward from TFBS central point. Default 1000}

\item{cov_limit}{Max base depth. Default 1000}

\item{mapq}{Min quality of mapping reads. Default 0}

\item{threads}{Number of threads. Default 1}

\item{output_dir}{Directory to output results}

\item{max_regions}{Max number of TFBS to analyze. Default 100000}
}
\description{
This function takes a DATA.FRAME with TFBS positions, a BAM file with the sequence to analyze, a mean genome wide coverage,
local coverage values for CNA, and outputs the corrected mean depth coverage values around TFBS and
a tfbs file with them. For better understanding check: https://www.nature.com/articles/s41467-019-12714-4
}
