% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{calculate_genowide_coverage}
\alias{calculate_genowide_coverage}
\title{Calculate Genome-wide coverage for a BAM file}
\usage{
calculate_genowide_coverage(
  bin_path = "tools/bedtools2/bin/bedtools",
  bam = "",
  verbose = FALSE,
  output_dir = ""
)
}
\arguments{
\item{bin_path}{Path to binary. Default tools/bedtools2/bin/bedtools}

\item{bam}{Path to BAM file.}

\item{verbose}{Enables progress messages. Default False.}

\item{output_dir}{Directory to output results.}
}
\description{
This function takes a BAM file and generates a TXT file with its genome-wide coverage
}
