% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.R
\name{calculate_MR_tfbs}
\alias{calculate_MR_tfbs}
\title{Calculate Mean Methylation Ratio around TFBS}
\usage{
calculate_MR_tfbs(
  bin_path = "tools/PileOMeth/output/MethylDackel",
  ref_data = "",
  bam = "",
  sample_name = "",
  tf_name = "",
  ref_genome = "",
  tfbs_start = 1000,
  tfbs_end = 1000,
  mapq = 10,
  phred = 5,
  output_dir = "",
  keep_strand = TRUE,
  bin_width = 50,
  verbose = FALSE,
  threads = 1
)
}
\arguments{
\item{bin_path}{Path to binary. Default tools/samtools/samtools}

\item{ref_data}{Data from BED file as Data.frame}

\item{bam}{Path to BAM file.}

\item{sample_name}{Sample name}

\item{tf_name}{Transcription Factor name}

\item{ref_genome}{Path to reference genome FA file.}

\item{tfbs_start}{Number of bases to analyze forward from TFBS central point. Default 1000}

\item{tfbs_end}{Number of bases to analyze  backward from TFBS central point. Default 1000}

\item{mapq}{Min quality of mapping reads. Default 10}

\item{phred}{Min phred base quality. Default 5}

\item{output_dir}{Directory to output results. If not provided then outputs in current directory}

\item{keep_strand}{Use strand information from BED files if available. Default TRUE.}

\item{bin_width}{Width of the the bins in which to group methylation data. Default 50.}

\item{verbose}{Enables progress messages. Default FALSE}

\item{threads}{Number of threads to use. Default 1}
}
\value{
A list of DATA.FRAMEs with the methylation ratio per base and per bin_width with default bin_width 50
}
\description{
This function takes a DATA.FRAME with TFBS positions, a BAM file with the sequence to analyze
and generates the values of mean Methylation Ratios per base around TFBS which then outputs into
a TSS file. For better understanding check: https://www.nature.com/articles/s41467-019-12714-4
}
