% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{char2seed}
\alias{char2seed}
\title{Generate seed from non-symbolic characters}
\usage{
char2seed(x)
}
\arguments{
\item{x}{A string}
}
\value{
A number
}
\description{
This function takes a string as input and transforms non-symbolic characters into a number
}
