% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{find_peaks}
\alias{find_peaks}
\title{Calls maximum and minimum peaks}
\usage{
find_peaks(x, m = 3)
}
\arguments{
\item{x}{A list with values}

\item{m}{Filter length}
}
\value{
A list with peak values
}
\description{
This function takes a list with a series of values and calls the minimum and maximum
}
