% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{get_coverage_bed}
\alias{get_coverage_bed}
\title{Get coverage for target regions}
\usage{
get_coverage_bed(
  bin_path = "tools/samtools/samtools",
  bam = "",
  bed = "",
  mapq = 0,
  threads = 3,
  output_dir = ""
)
}
\arguments{
\item{bin_path}{Path to samtools. Default tools/samtools/samtools}

\item{bam}{Path to BAM file}

\item{bed}{Path to BED file}

\item{mapq}{Minimum mapping quality}

\item{threads}{Number of threads to use. Default 3}

\item{output_dir}{Directory to output results}
}
\value{
A double with the mean genome coverage
}
\description{
This function takes a  BAM and a BED file with target regions and estimate per base coverage for them
using samtools depth
}
