% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{get_high_signal}
\alias{get_high_signal}
\title{Filters high signal frequency in TFBS coverage data}
\usage{
get_high_signal(coverage_signal, low_signal, m = 51)
}
\arguments{
\item{coverage_signal}{A list with coverage data}

\item{low_signal}{A list with the low frequency signal values}

\item{m}{Filter length}
}
\value{
A list with high frequency signal values
}
\description{
This function takes a list with the coverage data and normalizes and filters it leaving only the high frequency signal
}
