% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{get_low_signal}
\alias{get_low_signal}
\title{Filters low signal frequency in TFBS coverage data}
\usage{
get_low_signal(coverage_signal, m = 1001)
}
\arguments{
\item{coverage_signal}{A list with coverage data}

\item{m}{Filter length}
}
\value{
A list with low frequency signal values
}
\description{
This function takes a list with the coverage data and normalizes and filters it leaving only the low frequency signal
}
