% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{get_mean_and_conf_intervals}
\alias{get_mean_and_conf_intervals}
\title{Calculate mean and confidence intervals for positions relative to TFBSs}
\usage{
get_mean_and_conf_intervals(data = "", CI = 0.95)
}
\arguments{
\item{data}{LIST of DATA.FRAMES with the data}

\item{CI}{Confidence Interval}
}
\description{
This function takes a LIST of DATA.FRAMES as input(each DATA.FRAME for each TFBS), and returns a DATA.FRAME
with the mean and confidence intervals for the positions relative to them
}
