% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{get_mean_coverage}
\alias{get_mean_coverage}
\title{Calculate mean genome coverage}
\usage{
get_mean_coverage(
  file = "",
  output_dir = "",
  region = "genome",
  sample_name = "",
  save = TRUE
)
}
\arguments{
\item{file}{Path to file with genome-wide coverage}

\item{output_dir}{Directory to output results}

\item{region}{Region for which to get mean coverage. Default genome}

\item{sample_name}{Sample name}

\item{save}{Save as TXT. Default TRUE}
}
\value{
A double with the mean genome coverage
}
\description{
This function takes a TXT file with genome-wide ocverage as input and returns the mean genome coverage
}
