% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{get_norm_local_coverage}
\alias{get_norm_local_coverage}
\title{Get normalized local coverage from segmentation data (CNA) for a single base position}
\usage{
get_norm_local_coverage(pos = "", chr = "", norm_log2 = "")
}
\arguments{
\item{pos}{Genomic position}

\item{chr}{Chromosome}

\item{norm_log2}{Data.frame with normalized local coverage}
}
\description{
This function takes normalized data for local coverage as input,
as well as a valid genomic position for a single base and returns a log2 corrected value for said position
}
