% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{methylation_score}
\alias{methylation_score}
\title{Estimates the unranked Methylation score for TFBS}
\usage{
methylation_score(data = "", output_dir = "", verbose = FALSE)
}
\arguments{
\item{data}{DATA.FRAME with data or Path to TXT file}

\item{output_dir}{Directory to output results. If not provided then outputs in current directory}

\item{verbose}{Enables progress messages. Default FALSE}
}
\description{
This function takes a DATA.FRAME/ TXT file with the TFBS methylation ratio data and
estimates the Methylation score as the range between the global maximum and minimum of
the binned_data signal.
}
