% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_motif_MR.R
\name{plot_motif_MR}
\alias{plot_motif_MR}
\title{Plot Mean Methylation Ratio for TFBS}
\usage{
plot_motif_MR(
  data = "",
  trend_line = TRUE,
  tf_name = "",
  sample_name = "",
  output_dir = "",
  line = TRUE
)
}
\arguments{
\item{data}{Data.frame with mean depth coverage data.}

\item{trend_line}{Draw trend line. Default TRUE}

\item{tf_name}{Transcription Factor name}

\item{sample_name}{Sample name}

\item{output_dir}{Directory to output results.}

\item{line}{Draw a line between points. Default TRUE}
}
\description{
This function takes a DATA.FRAME with Mean Methylation Ratio data as input, and plots it
}
