% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{rank_accessibility}
\alias{rank_accessibility}
\title{Ranks the Accessibility Score for all the TFs}
\usage{
rank_accessibility(data = "", output_dir = "", verbose = FALSE)
}
\arguments{
\item{data}{Path to TXT file}

\item{output_dir}{Directory to output results. If not provided then outputs in current directory}

\item{verbose}{Enables progress messages. Default FALSE}
}
\description{
This function takes a path to a TXT file with the accessibility scores of all the TFs analyzed and ranks them.
}
