% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tmp_bed}
\alias{tmp_bed}
\title{Generates a bed file with temp suffix}
\usage{
tmp_bed(chr = "", start = "", end = "", strand = "", name = "File")
}
\arguments{
\item{chr}{A vector with chromosome names}

\item{start}{A vector with starting position}

\item{end}{A vector with end position}

\item{strand}{A vector with strand information}

\item{name}{A string with file name}
}
\value{

}
\description{
Generates a bed file with temp suffix
}
