% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze_MR_tfbs_around_position}
\alias{analyze_MR_tfbs_around_position}
\title{Analyze Mean Methylation Ratio around TFBS}
\usage{
analyze_MR_tfbs_around_position(
  bin_path = "tools/samtools/samtools",
  bin_path2 = "tools/PileOMeth/output/MethylDackel",
  bed = "",
  bam = "",
  tfbs_start = 1000,
  tfbs_end = 1000,
  ref_genome = "",
  max_regions = 1e+05,
  mapq = 10,
  phred = 5,
  verbose = FALSE,
  output_dir = "",
  plot = TRUE,
  keep_strand = TRUE,
  bin_width = 50,
  threads = 1,
  tfbs_data_keep = FALSE
)
}
\arguments{
\item{bin_path}{Path to binary. Default tools/samtools/samtools}

\item{bin_path2}{Path to secondary binary. Default tools/PileOMeth/output/MethylDackel}

\item{bed}{Path to BED file.}

\item{bam}{Path to BAM file.}

\item{tfbs_start}{Number of bases to analyze forward from TFBS central point. Default 1000}

\item{tfbs_end}{Number of bases to analyze  backward from TFBS central point. Default 1000}

\item{ref_genome}{Path to reference genome FA file.}

\item{max_regions}{Max number of TFBS to analyze. Default 100000}

\item{mapq}{Min quality of mapping reads. Default 10}

\item{phred}{Min phred quality. Default 5}

\item{verbose}{Enables progress messages. Default FALSE}

\item{output_dir}{Directory to output results. If not provided then outputs in current directory}

\item{plot}{Create a plot with coverage data. Default TRUE.}

\item{keep_strand}{Use strand information from BED files if available. Default TRUE.}

\item{bin_width}{Width of the the bins in which to group methylation data. Default 50.}

\item{threads}{Number of threads to use. Default 1.}

\item{tfbs_data_keep}{Keep unprocessed MR tfbs data. Default FALSE}
}
\value{
A DATA.FRAME with coverage data
}
\description{
This function takes a BED file with TFBS, a BAM file with an aligned sequence
and outputs mean methylation ratio values around TFBS, as well as
generates a PDF with a plot for them.
For better understanding check: https://www.nature.com/articles/s41467-019-12714-4
}
