% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{calculate_coverage_around_gp}
\alias{calculate_coverage_around_gp}
\title{Calculate Coverage around Genomic Position}
\usage{
calculate_coverage_around_gp(
  bin_path = "tools/samtools/samtools",
  chr = "",
  position = "",
  strand = "",
  bam = "",
  norm_log2 = 1,
  start = 1000,
  end = 1000,
  mean_cov = 1,
  mapq = 0,
  method = "default",
  start_bin = 75,
  end_bin = 75,
  score = "ACC",
  bin_width = 50,
  keep_strand = FALSE
)
}
\arguments{
\item{bin_path}{Path to binary. Default tools/samtools/samtools}

\item{chr}{Chromosome to analyze.}

\item{position}{Genomic position within the chromosome}

\item{strand}{Strand direction. If not available asumes + strand.}

\item{bam}{Path to BAM file.}

\item{norm_log2}{Local segment log2}

\item{start}{Downstream distance from genomic position to estimate coverage}

\item{end}{Upstream distance from genomic position to estimate coverage}

\item{mean_cov}{Mean genome wide coverage.}

\item{mapq}{Min quality of mapping reads. Default 0}

\item{method}{Method to use to estimate coverage. default/binned}

\item{start_bin}{Downstream distance from genomic position to limit bin size of central region. Only if binned mode is selected.}

\item{end_bin}{Upstream distance from genomic position to limit bin size of central region. Only if binned mode is selected.}

\item{score}{Score to measure. ACC/MR}

\item{bin_width}{Bin size to bin methylation data in default mode}
}
\value{
A DATA.FRAME with per base coverage por each genomic position
}
\description{
This function takes chromosome location, as well as strand information, and generates a coverage report around
for it. This coverage can be normalized if local segmentation values for the region are given and further corrected by the
the local coverage in the BAM
}
