% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_freq_decomposition.R
\name{plot_freq_decomposition}
\alias{plot_freq_decomposition}
\title{Plot frequency decomposition for TFBS coverage data}
\usage{
plot_freq_decomposition(data = "", output_dir = "")
}
\arguments{
\item{data}{LIST with data or Path to TXT file with it}

\item{output_dir}{Directory to output results.  If not provided then outputs in current directory}
}
\description{
This function takes a DATA.FRAME/ TXT file with the frequency data of the TFBS coverage and
generates a plot for it.
}
